requirements.txt
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blinker==1.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
entrypoints==0.4
executing==0.8.3
filelock==3.15.4
Fiona==1.8.22
Flask==3.0.3
fonttools==4.29.1
geopandas==0.12.1
idna==3.3
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.3.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
mysql-connector-python==8.0.31
nbclassic==0.3.6
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
notebook-shim==0.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyproj==3.4.0
pyrsistent==0.18.1
pyshp==2.3.0
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sklearn-pandas==2.2.0
sniffio==1.2.0
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
virtualenv==20.26.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==3.0.3
